add secret field and filter accounts
