+++ apps/pw2/app/accounts/types.ts
@@ -1,0 +1,7 @@
+export interface Account {
+	id: string
+	userId: string
+	website: string
+	username: string
+	password: string
+}
